requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Pillow==8.4.0
requests==2.26.0
